backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
scikit-learn
joblib
pandas

